+++ src/main/java/com/napier/sem/Country.java
@@ -24,7 +24,7 @@ public class Country {
         this.name = name;
     }
 
-    public Country(String name, int population)
+    public Country(String name, long population)
     {
         this.name = name;
         this.population = population;
